Random class impelemented

+++ src/fireAlarms/SensorInterface.java
@@ -3,7 +3,7 @@ package fireAlarms;
 //Sensor interface for every available reading sensors;
 interface  SensorInterface {
 
-	public int getTemperature();
+	public double getTemperature();
 	public int getBatteryLevel();
 	public int getSmokeLevel();
 	public int getCo2Level();
